+++ packages/client/lib/client/index.ts
@@ -155,7 +155,8 @@ export default class RedisClient<
 
     return async function (this: ProxyClient, ...args: Array<unknown>) {
       if (command.parseCommand) {
-        const parser = this._self.#newCommandParser(resp);
+        const parser = new BasicCommandParser(resp);
+
         command.parseCommand(parser, ...args);
 
         return this.executeCommand(parser, this._commandOptions, transformReply);
@@ -174,7 +175,7 @@ export default class RedisClient<
 
     return async function (this: NamespaceProxyClient, ...args: Array<unknown>) {
       if (command.parseCommand) {
-        const parser = this._self._self.#newCommandParser(resp);
+        const parser = new BasicCommandParser(resp);
         command.parseCommand(parser, ...args);
 
         return this._self.executeCommand(parser, this._self._commandOptions, transformReply);
@@ -194,7 +195,7 @@ export default class RedisClient<
 
     return async function (this: NamespaceProxyClient, ...args: Array<unknown>) {
       if (fn.parseCommand) {
-        const parser = this._self._self.#newCommandParser(resp);
+        const parser = new BasicCommandParser(resp);
         parser.pushVariadic(prefix);
         fn.parseCommand(parser, ...args);
 
@@ -218,7 +219,7 @@ export default class RedisClient<
 
     return async function (this: ProxyClient, ...args: Array<unknown>) {
       if (script.parseCommand) {
-        const parser = this._self.#newCommandParser(resp);
+        const parser = new BasicCommandParser(resp);
         parser.pushVariadic(prefix);
         script.parseCommand(parser, ...args);
 
@@ -342,10 +343,6 @@ export default class RedisClient<
 
   setDirtyWatch(msg: string) {
     this._self.#dirtyWatch = msg;
-  }
-
-  #newCommandParser(resp: RespVersions): CommandParser {
-    return new BasicCommandParser(resp);
   }
 
   constructor(options?: RedisClientOptions<M, F, S, RESP, TYPE_MAPPING>) {
